+++ resources/views/mail/admin/repair/repair-completed.blade.php
@@ -1,0 +1,15 @@
+<x-mail::message>
+# Repair Completed
+
+Hi, {{ $owner->name }}
+
+the services of your vehicle (license plate: {{ $car->number_plate }}) has been completed, please visit us again to check your vehicle and follow next
+administration process. Also let us know when there is something wrong with your vehicle after we do our service.
+
+
+
+
+best regards,
+
+Admin CAR WORKSHOP
+</x-mail::message>